Add new log level: TRACE (5)
